split lazy-loading backend into dedicated modules
